KAD-AD-99: move HealthIndicator-UnitTests to main package
